refactor: ethereum module and listener.
