+++ src/common/modules/ethereum/index.ts
@@ -1,0 +1,2 @@
+export * from './ethereum.module';
+export * from './ethereum.service';
