+++ src/endpoints/ethereum-indexed-data/ethereum-indexed-data.module.ts
@@ -3,13 +3,11 @@ import { ElasticsearchModule } from '@nestjs/elasticsearch';
 import { ServiceRequestController } from './service-request/service-request.controller';
 import { ServiceRequestService } from './service-request/service-request.service';
 import { LocationModule } from '../../endpoints/location/location.module';
-import { EthereumModule } from '../../endpoints/ethereum/ethereum.module';
 import { DebioConversionModule } from '../../common/modules/debio-conversion/debio-conversion.module';
 
 @Module({
   imports: [
     forwardRef(() => LocationModule),
-    forwardRef(() => EthereumModule),
     DebioConversionModule,
     ElasticsearchModule.registerAsync({
       useFactory: async () => ({
